// src/apispec/src/models/FileContentResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sandbox0 API
 * Public HTTP APIs exposed by internal-gateway.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface FileContentResponse
 */
export interface FileContentResponse {
    /**
     * 
     * @type {string}
     * @memberof FileContentResponse
     */
    content?: string;
    /**
     * 
     * @type {string}
     * @memberof FileContentResponse
     */
    encoding?: FileContentResponseEncodingEnum;
}


/**
 * @export
 */
export const FileContentResponseEncodingEnum = {
    Base64: 'base64'
} as const;
export type FileContentResponseEncodingEnum = typeof FileContentResponseEncodingEnum[keyof typeof FileContentResponseEncodingEnum];


/**
 * Check if a given object implements the FileContentResponse interface.
 */
export function instanceOfFileContentResponse(value: object): value is FileContentResponse {
    return true;
}

export function FileContentResponseFromJSON(json: any): FileContentResponse {
    return FileContentResponseFromJSONTyped(json, false);
}

export function FileContentResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): FileContentResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'content': json['content'] == null ? undefined : json['content'],
        'encoding': json['encoding'] == null ? undefined : json['encoding'],
    };
}

export function FileContentResponseToJSON(json: any): FileContentResponse {
    return FileContentResponseToJSONTyped(json, false);
}

export function FileContentResponseToJSONTyped(value?: FileContentResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'content': value['content'],
        'encoding': value['encoding'],
    };
}