// src/apispec/src/models/MoveFileRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sandbox0 API
 * Public HTTP APIs exposed by internal-gateway.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MoveFileRequest
 */
export interface MoveFileRequest {
    /**
     * 
     * @type {string}
     * @memberof MoveFileRequest
     */
    source: string;
    /**
     * 
     * @type {string}
     * @memberof MoveFileRequest
     */
    destination: string;
}

/**
 * Check if a given object implements the MoveFileRequest interface.
 */
export function instanceOfMoveFileRequest(value: object): value is MoveFileRequest {
    if (!('source' in value) || value['source'] === undefined) return false;
    if (!('destination' in value) || value['destination'] === undefined) return false;
    return true;
}

export function MoveFileRequestFromJSON(json: any): MoveFileRequest {
    return MoveFileRequestFromJSONTyped(json, false);
}

export function MoveFileRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): MoveFileRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'source': json['source'],
        'destination': json['destination'],
    };
}

export function MoveFileRequestToJSON(json: any): MoveFileRequest {
    return MoveFileRequestToJSONTyped(json, false);
}

export function MoveFileRequestToJSONTyped(value?: MoveFileRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'source': value['source'],
        'destination': value['destination'],
    };
}