// src/apispec/src/models/SuccessFileReadResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sandbox0 API
 * Public HTTP APIs exposed by internal-gateway.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { SuccessFileReadResponseAllOfData } from './SuccessFileReadResponseAllOfData';
import {
    SuccessFileReadResponseAllOfDataFromJSON,
    SuccessFileReadResponseAllOfDataFromJSONTyped,
    SuccessFileReadResponseAllOfDataToJSON,
    SuccessFileReadResponseAllOfDataToJSONTyped,
} from './SuccessFileReadResponseAllOfData';

/**
 * 
 * @export
 * @interface SuccessFileReadResponse
 */
export interface SuccessFileReadResponse {
    /**
     * 
     * @type {boolean}
     * @memberof SuccessFileReadResponse
     */
    success: boolean;
    /**
     * 
     * @type {SuccessFileReadResponseAllOfData}
     * @memberof SuccessFileReadResponse
     */
    data?: SuccessFileReadResponseAllOfData;
}

/**
 * Check if a given object implements the SuccessFileReadResponse interface.
 */
export function instanceOfSuccessFileReadResponse(value: object): value is SuccessFileReadResponse {
    if (!('success' in value) || value['success'] === undefined) return false;
    return true;
}

export function SuccessFileReadResponseFromJSON(json: any): SuccessFileReadResponse {
    return SuccessFileReadResponseFromJSONTyped(json, false);
}

export function SuccessFileReadResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): SuccessFileReadResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'success': json['success'],
        'data': json['data'] == null ? undefined : SuccessFileReadResponseAllOfDataFromJSON(json['data']),
    };
}

export function SuccessFileReadResponseToJSON(json: any): SuccessFileReadResponse {
    return SuccessFileReadResponseToJSONTyped(json, false);
}

export function SuccessFileReadResponseToJSONTyped(value?: SuccessFileReadResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'success': value['success'],
        'data': SuccessFileReadResponseAllOfDataToJSON(value['data']),
    };
}